requirements.txt
deta<1.1.0
python-dotenv==1.0.0
requests==2.28.1
streamlit==1.20.0
streamlit_authenticator==0.1.5
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
